Move comment finding logic to jsdoc
